api/requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
influxdb-client
httpx
python-dotenv
